refactor(v4): file smoothTo under utils, where v3 had it
I had put it at `src/` root, reasoning that a util importing a service is a
layering inversion and pointing at `viewTransition.ts` as the precedent. That
argument does not beat where the thing belongs: v3 shipped it as
`utils/math/smoothTo.ts` and it is a util.

Being a util it leaves the root barrel — utils are not exported from
`src/index.ts` — and joins `src/utils/index.ts` instead.

Worth knowing that this makes the utils barrel reach into the services graph:
importing `utils/index.js` now pulls the raf service and the scheduler with it,
where before it was four leaf modules. Deep imports avoid that, and are what
`src/` uses internally.
